requirements.txt
Django==3.1.1
django-ckeditor==6.2.0
django-crispy-forms
django-jazzmin==2.5.0
django-phonenumber-field
django-rest-framework
django-settings-export
phonenumbers
Pillow==8.4.0
crispy-bootstrap5==0.6
six==1.16.0
django-multiselectfield